bitbake: hashserv: Add request statistics
Adds support to the bitbake hash server for recording statistics. Two
sets of statistics are reported: The number of connections and how long
until the requests started to be serviced on them, and the number of
requests and how long each request took to be serviced.

Statistics are available via an HTTP end point on the server:

 curl http://server:8686/v1/stats | jq

The statistics can be reset by issued an HTTP DELETE request to that end
point:

 curl -X DELETE http://server:8686/v1/stats

Signed-off-by: Joshua Watt <JPEWhacker@gmail.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>