bitbake: hashserv: Add more threads
Adds support for running the hash server with more threads. If the
server is invoked automatically by bitbake it will use a single thread
because in the presence of a single client this is the fastest
mechanism.

If the server is started manually on the command line, it is assumed
that multiple clients will be connecting, so it will use one thread per
CPU core (although since the server might be I/O bound not CPU bound,
this still may not be sufficient for optimal performance).

Signed-off-by: Joshua Watt <JPEWhacker@gmail.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>